Move enums in package
