+++ baremaps-core/src/main/java/org/apache/baremaps/openstreetmap/model/ChangeType.java
@@ -12,18 +12,6 @@
 
 package org.apache.baremaps.openstreetmap.model;
 
-
-
-import java.util.List;
-import java.util.StringJoiner;
-
-/** Represents a change in an OpenStreetMap dataset. */
-public record Change(ChangeType type,List<Entity>entities){
-
-  /** {@inheritDoc} */
-  @Override
-  public String toString() {
-    return new StringJoiner(", ", Change.class.getSimpleName() + "[", "]").add("type=" + type)
-        .add("elements=" + entities).toString();
-  }
+public enum ChangeType {
+  DELETE, CREATE, MODIFY;
 }
