+++ baremaps-core/src/main/java/org/apache/baremaps/openstreetmap/model/Element.java
@@ -24,6 +24,7 @@ import org.locationtech.jts.geom.Geometry;
  * world.
  */
 public sealed
+
 abstract class Element implements Entity
 permits Node, Way, Relation
 {
